:recycle: [Attributes] Move out entities from Annotations namespace
