+++ config/packages/doctrine.yaml
@@ -21,12 +21,6 @@ doctrine:
                 dir: '%kernel.project_dir%/src/Entity/Attributes'
                 prefix: 'App\Entity\Attributes'
                 alias: AppAttributes
-            AppAnnotations:
-                is_bundle: false
-                type: annotation
-                dir: '%kernel.project_dir%/src/Entity/Annotations'
-                prefix: 'App\Entity\Annotations'
-                alias: AppAnnotations
             App:
                 is_bundle: false
                 type: yml
