+++ src/Repository/ClientPermissionRepository.php
@@ -2,7 +2,7 @@
 
 namespace App\Repository;
 
-use App\Entity\Annotations\ClientPermission;
+use App\Entity\Attributes\ClientPermission;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Persistence\ManagerRegistry;
