+++ src/RepositoryV2/ResetPasswordRequestRepository.php
@@ -2,7 +2,7 @@
 
 namespace App\RepositoryV2;
 
-use App\Entity\Annotations\ResetPasswordRequest;
+use App\Entity\Attributes\ResetPasswordRequest;
 use App\Entity\User;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Persistence\ManagerRegistry;
